Render gov nav if base_path has "/government"
When a email alert signup starts with "/government" in its base path, we
should be displaying the government navigation. This commit does that by
checking the `base_path` and getting the second section of this to mark
the active link in the nav. This follows the same approach as Finder
Frontend.