Use the email_alert_type when constructing the email-alert-api payload
This makes use of a new field on email_alert_signup content items -
'email_alert_type' - to construct the payload for the email-alert-api
when sending a subscription request. Using the field means the frontend
doesn't need to have explicit knowledge of the types of email alert that
are supported. Currently only policy subscriptions are handled in this
way, but the implementation leaves open the possibility of migrating
topics (for example) to make use of the email-alert-frontend as well.

Govuk content schemas determine which values of email_alert_type are valid,
ensuring that the email-alert-api receives expected values.