Remove non-working breadcrumb code
The breadcrumbs in this application seem to never have worked. The code
constructs a hash and passes this into slimmer as a "dummy artefact".
But this only works for pages that are rendered with the `core_layout `
layout from static, and these pages are rendered with `core_layout`.

This commit removes the code, since it's not working. It'll allow us to
upgrade to a new slimmer version.