Support subscriptions to associated taxons
Associated taxons (those that are included within another taxon e.g.
`/world/birth-death-and-marriage-abroad` that is included in all country
taxons are currently causing errors when users try to subscribe to them.

This is because, whilst they get a different slug and content id, their
title within the content item stays the same. Govdelivery currently
enforces and uniqueness constraint on subscription names and we use the
content item title for this purpose.

Consequently, when a user tries to subscribe to an associated taxon it will
fail if that title has already been used. Ideally, this constraint would
be removed at Govdelivery as it seems to serve no purpose given that the
subscription topics already have a unique ID with Govdelivery. As we are
unable to make this happen currently this commit favours the taxon
`'details' => {'internal_name' => <title>}` if it is present.