Add override task for email-signup routes
The base path /email-signup was previously registered by Content Tagger.
Email Alert Frontend should be the `publishing_app` and rendering app so
we need a task to force override the `publishing_app` in Publishing Api.

We can then use the tasks to register special routes for `email-signup`
and `email-signup/confirm` without them erroring.