Update styling on unsubscription pages
Add GOV.UK header component on both confirm and confirmed pages with
some context and a title.

Remove the grid styling on the confirm page to bring text and button
into correct alignment.

Change the submit button on the confirm page from an `input
type="submit"` tag to a `button` tag to honour the GOV.UK Static
Component Guide.