Remove the condition around exposing the /email/subscriptions path
Graham raises a good point in the pull request for this branch that we
shouldn't expose this path to members of the public before we're ready
to switch over to using it.

We actually already get this for 'free' because we have to manually
publish a prefix route in production for this to happen:
`rake publishing_api:publish_email_subscriptions_prefix`

Therefore, I think it makes sense to remove this condition and keep the
routes file a bit simpler.