Update subscribe feature tests for frequency
This updates both the signup test and back tests.

For the signup one it's been relatively simply changed to just have the
extra step.

The back button seemed to test a few scenarios in a single test so I've
broken those down to individual contexts. This didn't really feel ideal
for a feature test and felt like it should be somewhere else, however it
is easier to test in this context than as a controller test.