Set GOVUK_ASSET_ROOT dynamically for heroku apps
Since we'll be using the Procfile in more cases than just Heroku it no
longer makes sense to always set GOVUK_ASSET_ROOT there.

I tried exploring a few options of setting the value as part of a
postdeploy script but that didn't seem to work so it seemed easier to
just not set that env var for Heroku deploys and then set it
dynamically. If we want it to be something else we can always set the
env var (which wasn't previously an option).