Revert app to not use foreman in startup.sh
There are still some problems to resolve running the app via unicorn in
a dev environment and these may take a while to be resolved. So this
reverts this back to the previous dev experience of starting rails
server directly. This can then be switched around later if the problems
are resolved.