Render the frequency question in a legend
This uses a govspeak component inside the legend of the form to inherit
govspeak styles. We also apply a class to the header so that we can give
this title a bottom margin.

This also drops the second line "You can get updates about changes:" as
having this in the legend made for a confusing screenreader experience.