Clarify which JS requires come from govuk_frontend_toolkit
In general we don't want to require all of the JS from
govuk_frontend_toolkit, because in deployed environments we'd end up
duplicating a lot of the JS that we load in from static.  However,
depending on which static layout we use we don't actually get all of
govuk_frontend_toolkit's JS, and we might want to rely on parts that
aren't delivered by static.  In these cases it's fine to require those
files, but we should be clear about what we can and can't require to
avoid someone changing the require to `//= require govuk_toolkit` in
the future.

To that end we drop a comment at the top of `application.js` around
the existing requires explaining that they are parts of
govuk_frontend_toolkit that are not included by the `header-footer-only`
template we use.