Set Cache-Control header and allow CSRF protection
This commit sets the `Cache-Control` header for all controllers that serve pages under `/email` to “private” so they don’t get cached by the CDN. This is because cookies and sessions will be enabled soon and we don’t want specific users’ pages to be cached. It also enables CSRF protection now that cookies are enabled.