Determine unsubscribe back links based on authentication presence
Rather than using a query string to determine if they were from
subscription management this instead relies on whether they are
authenticated or not.

There is of course the situation that they could be authenticated and
have gone to this link via email, but in that situation it's probably a
better experience to end up in your subscription management.