Don't use internal_name of taxons
The internal_name is not meant for public consumption, it's used in
content tagger to give more context to the taxonomists.

This makes it so the subscriptions are created with just the title.

The internal_name was introduced in
https://github.com/alphagov/email-alert-frontend/commit/78fdf94e64f61c1f
86bb91fc4aadfc5ff602aa65.

The reasoning behind it is that the titles needed to be unique in the
system, something that was enforced by GovDelivery. Because we're no
longer using GovDelivery, we should be able to remove the uniqueness
constraint.