Revert "Support subscriptions to associated taxons"
Internal names are by definition internal, so we shouldn't be using
these as user visible titles for subscriptions.

The original motivation for this related to GovDelivery, which is no
longer in use.

As far as I understand, there looks to be a general issue with email
subscription lists in the email-alert-api service having a unique
constraint on the title, which will probably cause some issues with
Topic Taxonomy subscriptions as the titles of taxons are not unique.

This reverts commit 78fdf94e64f61c1f86bb91fc4aadfc5ff602aa65.