Ignore a false positive about URL redirects
We could find a way in the code by parsing the URI, extracting
the various parameters and then passing them individually to
`redirect_to` but with an explicit `host` parameter. But even if
we did that, the host would come from Plek and Brakeman would
complain about that. Plus we already trust the data in the
Content Store.

Instead, it seems easier to ignore this warning.