Make sure signup page can handle all valid taxons
We had a small incident this morning where taxonomy signup pages (like
https://www.gov.uk/email-signup/?topic=/education) were errorring
because the taxon didn't have a description. It was fixed in
https://github.com/alphagov/email-alert-frontend/pull/399.

The way to prevent these problems is to test the page against random
content items. In this case it would have prevented the problem because
`description` isn't required, so a random content item generated by
`GovukSchemas::RandomExample` is likely to sometimes return a taxon
without the description. This would've triggered a test fail.