Generalise TaxonomySignup model This replaces the TaxonomySignup model with a new model, ContentItemSubscriberList.
End users wont see a change as a result of this commit.
The intent is to make it possible to give an arbitrary
content item to the ContentItemSubscriberList and have it
return a subscriber list for that content item.

TaxonomySignup does this just for taxons. Now we can use
ContentItemSubscriberList to generate subscriber lists
for people, organisations, and so on.

In a future commit I will use this class to generate
subscriber lists for organisations.