Consolidate all subscription frequencies in i18n locale
Over time, the repository had copied & pasted the same bits of text in various places, making it difficult to update.
For example, updating 'No more than once a day' to 'No more than once daily' would involve changing in 3 places.

If we were to add additional frequencies, e.g. fortnightly, again we'd need to touch multiple files.

All of this information is now kept in a single config/locales/en.yml, so that vocabulary can be changed and frequencies edited in one place.