Implements the new unsubscribe logic
Currently, if a user changes their subscription frequency and then
clicks 'Unsubscribe' in one of their older emails, they'll see a page
saying "You're already unsubscribed". This is confusing because the
user hadn't really unsubscribed in the first place; the fact that
our back end treats a frequency change as a new subscription is not
something the user should know or care about. Moreover, this message
might lead the user to believe they have now unsubscribed from the
subscriber list, when in fact they are still subscribed.

To fix this, we now attempt to fetch the most recent subscription a
user has for a given `subscriber_list`, and respond appropriately:

- If their latest subscription has ended already, we'll show the
"You're already unsubscribed" page
- Otherwise, if this is their latest subscription, we'll render a
"Are you sure you want to unsubscribe?" page
- Otherwise, if this is not their latest subscription, but its
frequency matches that of their latest subscription, then we can
act on the user's intention and unsubscribe them (from their latest
subscription). They don't need to know it has a different ID.
- Or if this is an older subscription and its frequency is different
to that of their latest subscription, we cannot act on their instruction
to unsubscribe as it may not be what the user intended. Instead we
redirect them to manage their subscription manually.

Trello card: https://trello.com/c/VlnvU2Dl