Create SubscriberList for other doc types
This commit allows other document types to be used to create subscriber
lists in email-alert-api. There is some weirdness here when it comes to
document with a type of 'taxon' as this could either be an actual taxon
so '/education' or a 'world location' so '/world/france', this is even more
confusing as a document can have a type of world_location which is actually a
international delegation.

The links on content for international delegations
(https://www.gov.uk/world/uk-joint-delegation-to-nato) and 'world
locations' (https://www.gov.uk/world/france) both have a key of
`world_locations`.

World locations (https://www.gov.uk/world/france) are not treated like
taxons even though they are technically and so need a seperate link
hash.

There have been a few cards on the gov.uk tech debt board relating
to these type of issues.