Amend back button behaviour
This addresses one signup edge case where if a user clicks through the
signup flow and lands on the address input page and then decides to
travel back using our back link, they will first arrive on the frequency
selection page then if they click back again they will be redirected to
gov.uk's homepage, this is obviously not a great user experiance and
affects any email signup flow.

It is also possible to get trapped in an infinate loop using the back
links between the confirm page and the frequency page for signups that
start in email-alert-frontend.

This commit targets one edge case regarding email subscriptions from
/get-ready-brexit-check/results as we now can access the subscriber list url to
find the original page the user signed up to. This is hacky and a solution
to fix the overriding problem will need to be thought of.