Rename AuthenticationController to be specific
This renames the AuthenticationController to reflect that it's specific
to the workflow for managing subscriptions:

   - It requires a subscriber to exist in order to create a session
   - It redirects to the management workflow if the token is invalid

The upcoming work to support double opt-in for subscriptions is not
compatible with this behaviour. Instead of trying to make the existing
controller more generic, we believe it will be simpler to implement a
new one with additional functionality. Hence renaming the existing one.