Deprecate top-level '/authenticate' routes
This moves the existing '/authenticate' routes under the '/manage' scope
to reflect that they are specific to the actions within and support the
upcoming work that will introduce additional authentication routes for
a new subscription (as opposed to managing subscriptions / loggin in).

We need to keep the two 'GET' routes to avoid breaking links in existing
emails. One route can be a redirect, but the other has query params,
which aren't easy to preserve in a redirect. However, we're confident we
can delete the latter after 7 days, since all emails that contain it
will have expired, and a 404 seems reasonable.