Remove legacy route to process sign in token
Previously we scoped this route under '/manage'. This removes the
old route, which we kept for backwards compatibility. New auth emails
should now be using the scoped URL, and since after 7 days all the
emails containing the old URL are now expired, we think it's reasonable
to completely remove the old route.