Rewrite email alerts feature test
This rewrites the Cucumber feature around defining a new subscriber list
via a '<base_path>/email-signup' route, as well as making the test more
clearly indicate that a subscription is *not* created (until later). A
tangent of the original feature test also checked for the presence of a
'proposition_menu', even though the rendered content was unchanged. This
is not carried over into the new feature spec, as it seems redundant.