Add new opt-in subscribe feature
This adds a new feature to create (or update) a subscription against a
validated JWT token. Since the address to subscribe to is encrypted
within the token, we can be sure the request represents genuine intent.
As an additional protection, we also validate the subscriber list
matches the one in the token, so that it can't be used out of context.