Follow redirects for content item email signups
Sometimes taxa are moved and a redirect is set in place.
Email signup links to the moved taxons are often hardcoded in content
and will have to be updated. Some may be missed resulting in broken links.

Previously, the case of redirecting /government/brexit -> /brexit was handled
as a special hard coded case.

This PR makes makes this more general by following the redirects in
the content item.