Show error when user tries to confirm email-signup without a topic
Previously, we redirected to root (GOVUK homepage) for all instances when requests to new/confirm/create had invalid or missing params.
There are cases where we may still want to do this, so I've tried to keep those as-is.

However, if a user tries to click through from the initial taxon sign-up page without selecting a topic, we now show a useful error message and redirect them back to the selection page.