Support content item signup for service manual
Previously these relied on the 'email_alert_signup' functionality
in this app, which we'd like to remove, since most formats now use
the generic 'content_item_signup' functionality. This adds support
for creating subscriptions based on the content item.

Note that the single instance of 'service_manual_service_standard'
is linked to by a somewhat anonymous 'parent' key, instead of the
normal, pluralised keys for multi-instance formats, like taxons.