DRY-up tests for ContentItemSubscriberList
Previously we were repeating the same expectations over and over,
even though only one output changes between the contexts. This
removes the unnecessary repetition, noting that it's still necessary
to call 'subscription_management_url' in order to trigger the API
call under test. We will remove this oddity in a future commit.