Stop passing invalid things around
Previously the controller passed a potentially nil content item to
the subscriber list wrapper class, but then explicitly asked the
class if the item was present. This replaces the unnecessary indirection
by making the controller do the check itself. Since the wrapper class
now knows the content item is valid, it can pre-cache the subscriber
list, to avoid multiple API calls in future.