Add extreme EmailVolume value
The only taxon we are getting extreme volumes for are notifications
for content tagged with the Coronavirus taxon. We want to inform
the user what these numbers are, so that they can get decide whether
they still want to get immediate notifications. The min value is
based on a daily average, and the maximum is on a day's maximum
content changes during the first couple of weeks of the taxon's
existence.