Extract verifying an email into services
As part of creating or managing a subscriptions, we need to send a
verification email so that the user can prove they own the address
to which the subscriptions pertain. This extracts the API calls to
send these verification emails into two new services, with the aim
of extending them to rate limit these actions.

Note that much of the code of this app is written inline with the
controllers. This adds a new 'ApplicationService' concept in order
to provide a place for extending the existing code, without cluttering
the controllers. Defining a service this way is based on the approach
we take in Content Publisher [1].

[1]: https://github.com/alphagov/content-publisher/issues/1256