Rate limit verification emails
This adds the 'ratelimit' gem used by email-alert-api and uses it
in the new services in order to prevent an excessive number of
requests to create or manage subscriptions on the same address; the
exact thresholds will be updated in a later commit.

Since the thresholds are relatively low, we need to globally stub
the rate limiter to avoid rate limiting our tests.

Note that the rate limiter uses a hard-coded Redis namespace [1],
so we need to encode the app name into the rate limit key in order
to avoid future clashes with other apps sharing the same Redis
instance.

[1]: https://github.com/ejfinneran/ratelimit/blob/0e60d3484b4b1f984f806b1c89c601adf92fd544/lib/ratelimit.rb#L118