Verify user owns subscription before updating
This fixes a problem where using the endpoint to change_frequency was
raising an error if a subscription was modified that didn't belong to a
particular subscriber.

It also updates the controller test for this endpoint. Both posts and
gets merely check whether a subscription belongs to a subscriber rather
than exists at all, since this case also covers non-existent
subscriptions.

I fixed the test for an invalid frequency, this was previously not
testing anything as it relied on a successful HTTP request mock. Instead
Email Alert API will return 422 which will be converted into a
BadRequest response by Email Alert Frontend.