Update check email screen with new design
This makes a few changes to accommodate the new design:

- The back link is only rendered in the error case, so the code to
set it is moved accordingly.

- The subscriber list is fetched in order to show the title.

- The CSS for the "details" component is added, since this is the
first time we're using it in this app.

I've also chosen to switch to plain HTML tags everywhere, instead
of "tag.<tag>", as this was inconsistent within the file, and the
avoidance of this seems to be the norm in other apps [1].

[1]: https://github.com/alphagov/content-publisher/blob/master/app/views/topics/edit.html.erb