Stop having "Back" links to absolute URLs
Previous we incorrectly and unnecessarily used "url_for" to generate
the URLs for "Back" links on various pages of the latter parts of the
signup flow. Using "url_for" means the resulting URL is unnecessarily
absolute, and, without a "host" parameter, it also incorrectly absorbs
any referrer. This switches to using more conventional path helpers in
place of "url_for", which we seldom test as "back" is not considered
to be critical functionality.