Rewrite content for development homepage
Since a developer may not think to spin up the app in order to
learn about it, we should favour having documentation in the
README, which happens to already cover all of the content here.
On the other hand, keeping this page still provides a nicer
experience for someone who's new to the app. This replaces the
previous content with some simple links to further info.