Avoid using list URLs from Email Alert API
Previously we redirected to the "subscription_url" returned by Email
Alert API in order to create a new subscription for a list, which is
already rendered by Email Alert Frontend. Originally this was done to
support migration from GovDelivery. Nowadays it just causes problems
with local development, since the returned URL is for "www.dev.gov.uk"
and not "email-alert-frontend.dev.gov.uk".  This switches out the URL
with an explicit redirect to the native path for new subscriptions.

[1]: https://github.com/alphagov/email-alert-api/commit/c39b4970ed8bcdfb47797117e7722af9477f73c5#diff-dbcd66db9f7631a8473bb321924a02419acd89199d38cd2e519982c8d0eb0ff6R10