DRY up knowledge about supported content types
Previously the ContentItemSignupsController had its own knowledge
about which types of content were supported, which was repeated in
the model that would find/create the associated list after generating
the associated query params. By redefining the model interface to
just do the params generation part, we can DRY up the knowledge of
which types of content are supported, with the controller doing the
list find/create part itself.