Condense getting signup content item into method
Previously we had three methods that had to be called in sequence
as part of getting a content item, with other code then usign a
mixture of the '@content_item' and 'content_item' afterwards. This
combines all the parts of the request handling related to fetching
a content item into a single 'assign_' before_action.