Clarify specs for content store request handling
Previously some of these made a hard-coded request using "get :new",
despite being shared examples. Some of the descriptions incorrectly
referred to "redirect", and the use of "taxon" was excessive for what
is a now a generic endpoint. This rewrites the descriptions to make
them accurate and consistent, and fixes the request handling so that
the shared examples work with all three routes under test.