Rewrite unsubscribe spec to test two scenarios
Previously we only tested the unauthenticated journey. This rewrites
the feature to cover the two scenarios where someone may unsubscribe:

- With an unauthenticated link in a digest email.

- Via the manage subscriptions page (after authenticating).

This is a precursor to changing the journey so that the user must be
authenticated when unsubscribing.

Note that previously we were incorrectly passing the ID of the list
as the ID of the subscription for one of the test stubs.