Allow a user to unsubscribe by list
Previously we added functionality to redirect if the subscription
was not the latest for a list e.g. if the user had previously ended
and then restarted their subscription [1]. This makes a few changes to
the original functionality:

- We no longer compare frequencies of the requested and latest
subscriptions. A user can only have one subscription per list, so
it seems unlikely that they will accidentally unsubscribe after
previously changing the frequency [2].

- The resulting functionality is refactored into a before_action, to
simplify the main action logic.

[1]: https://github.com/alphagov/email-alert-frontend/pull/474

[2]: "the fact that our back end treats a frequency change as a new
subscription is not something the user should know or care about"
(from [1])