Ignore if the user already unsubscribed
Previously this only affected the journey if the user was signed in,
yet only the non-signed-in path was tested (and the description of
the test was inconsistent with its assertions). Since the user likely
thinks they *do* have a subscription when they press the button, this
removes the unnecessary and untested conditional.