Require some kind of authentication to unsubscribe
Previously we allowed any user to unsubscribe any other user. This
fixes that to enforce that either:

- The user is signed in and owns the subscription; or
- The URL has a token that refers to the subscription to end

Note that for some of the existing specs, we now need to provide some
sort of authentication just to exercise the same code.

While the digest emails currently make use of the unauthenticated API,
we want to require authentication from now on, and the existing links
to "Unubscribe" in digest emails will now redirect to the sign in page.
Indeed, we're going to move away from one-click unsubscribe links in
digest emails, as a temporary, simplifying measure before we look at
having individual digest emails for each subscription.