Add redis connection healthcheck
This adds some added functionality to the healthcheck for
email-alert-frontend to enable CD.

As per the requirement for CD: 'It has a test that checks for
connectivity to any read/write dependencies'. The only DB
email-alert-frontend uses is Redis during [rate limiting]. Opening a
Redis connection and then adding, fetching and deleting a key value pair
before closing the collection seems a sufficient test for Redis
connectivity.

[rate limiting]: https://github.com/alphagov/email-alert-frontend/blob/a32156a677f3989623faf6953de6aa4c8e073e29/app/services/verify_subscriber_email_service.rb#L52