Record the origin of new content item-based lists
The URL property of subscriber lists was introduced to support the
Brexit Checker [1], but the functionality applies to all lists. We
think that recording these URLs will in future help us to determine
which lists are still "alive", by checking they refer to a live URL
on GOV.UK. This takes a step in that direction.

[1]: https://github.com/alphagov/finder-frontend/blob/2b565b6ccb6449d329dba00bb39d4d629a42c81a/app/controllers/brexit_checker_controller.rb#L189