Sign in after subscribing
Previously the user would be shown a dead-end "you've subscribed"
page after confirming a subscription. This changes the controller
to create a session, with the "you've subscribed" message being
shown in a flash on the manage subscriptions page.

Note that, unlike the subscriber authentication controller, we do
not destroy any existing session if the token is invalid or expired
[1], since the session is still valid - the change here is simply a
convenient alternative to the normal sign in.

Note also that, for simplicity, the subscriber subscriptions stub
in the feature test returns an empty array of subscription objects,
despite one having been created, since this is not relevant to the
code under test.

[1]: https://github.com/alphagov/email-alert-frontend/commit/41c38680354e6e2c2b98bbcf59c116d1ae909a81#diff-6966e9ac443db47e60dc97c2beb2864ce79ab9a4ce5e477d5280232620d4baedR38