Update flash message following design review
This updates the heading text to be consistent with the confirmation
email we send. It also updates the way the content of the flash box
is presented, in order to clearly distinguish the list subscribed to
from the content introducing it.

Since the current "success" flash only allows a single line of text,
this adds a new "subscription" flash type to highlight information
about a subscription. We will reuse this in subsequent work as the
confirmation for changing frequency.

Note that this introduces the first view test, in order to test the
rendering of different messages, which is coded as part of the view,
and would be excessive or irrelevant to test at a higher level, such
as a controller or feature test.