Defend against missing subscription in flash
Previously we would error if the "subscription" flash referred to
a subscription that was not present in the hash retrieved by the
controller. While this is unlikely, the tenuous link between the
subscription in the flash and the subscriptions retrieved by the
controller means we should defend against future uses where they
may not be in sync. Since the flash is only there to give optional
feedback, it makes sense to silently ignore any errors it causes.