Update login screen to match design
We will update the error flashes in a subsequent commit. Note that
the new "Email address" label also matches against the hidden one
for users to give feedback on a page. This switches to using the
name of the field to avoid the ambiguity, at the expense of being
less representative of the link with the field label.