Remove unnecessary <br> element in flash
This was only necessary because we didn't apply the "govuk-body"
style to the <p> element, which restores the normal spacing after
a <p> element (GOV.UK removes this by default).

Note the following paragraph does not have this class, as it would
introduce additional blank space at the end of the flash message.