Collect and revise technical documentation
This moves the existing, more technical, parts of the README into
the standard section we have for this. Note that, while this is a
frontend app, most of the features work only work in tandem with
the protected Email Alert APIs, so we should treat it like most of
our other apps and use GOV.UK Docker instead of a startup script.