Shorten and revise section on signup feature
This combines the explanation of each signup journey with the bullet
point declaring it. This also removes some of the detail around double
opt-in, as this wasn't particularly helpful:

- It was out-of-date (especially the part about the strategy changing)
- It's inconsistent: we don't document the login flow to manage emails.
- It's out-of-place: the documentation is equally relevant in the API.

If we find we need to document the login behaviour in future, then we
should do so more thoroughly and in a separate doc.