Add a controller method to authenticate a user by GOV.UK Account
We added a new endpoint to email-alert-api to authenticate a user by
GOV.UK Account, returning the subscriber details if the email address
is verified and matches a subscriber.  This commit makes the
corresponding change on the email-alert-frontend side, exposing that
feature to users.

This is not complete, we still need to solve some design issues, like:

- What do we show in the error cases?

- Do we prompt a user to explicitly "link" the accounts, rather than
just sign them in?

- Do we forbid users from changing their notifications email address
if it's linked to an account?  Would it be confusing if the addresses
get out of sync?

So for now it's hidden behind a feature flag.

See also https://github.com/alphagov/email-alert-api/pull/1638