Tell people to log in via the account if they have one
If a user goes straight to /email/manage/authenticate, we need to send
them on the right path: to the account if they have that, to the auth
token if they don't.

So I've added a bit of text to this page to try to get people to do
the right thing.  If they try to use the auth token when they have an
account, they'll see an error page telling them to sign in, due to the
previous commit.

This new text has not been content-designed, this is
good-enough-for-development text I have put together which we should
review before allowing users to link accounts.