Link subscriber to GOV.UK account on auth
We will probably NOT want to do this when we launch it: we've talked
about showing an interstitial page to tell the user that their
notifications are not managed through their account.  But this is good
enough for demonstration purposes.

When we do want to do an interstitial page, we'll instead want some
logic like this:

    subscriber = auth_with_govuk_account
    if subscriber.linked_to_govuk_account
        redirect_to list_subscriptions_path
    else
        render :welcome_to_the_brave_new_world

Where the new page has a "Continue" button it which then calls the
link endpoint.