Implement account layout for email management
The accounts team are working towards enabling users who have linked
their email notifications address and their GOVUK account to manage
their email subscriptions through their account.

Part of this work is making the email management pages in
email-alert-frontend look the same as the account when the user is
authenticated with a linked account.

This PR reorganises some of the views and requests the
`gem_layout_account_manager` template from static which will
automatically apply account page furniture to the email management pages
when the user is logged in with a linked account.