Create a things-are-changing page
Using placeholder content for now, so it is hidden behind a feature
flag, and may require extension.

This is a part of the project to more merge the GOV.UK notification
accounts and the new GOV.UK Account. As such we'll now tell users that
their existing accounts will need to be linked to GOV.UK account.

1. Move govuk_account_auth_enabled? out into a helper

If we move this to a module we can re-use it in another controller

Add /things-are-changing with placeholder content

Nests the page behind the GOV.UK Account Feature flag

Content to follow when confirmed

Also subsequent commit will add conditionality that varies the text of
the call to action button based on if a user is logged in or not.

/things-are-changing vary button by logged_in?

Ends up going to the same place as a logged out user will attempt to
make the authenticated request and be redirected to sign-in, but we'll
want to vary content to prevent confusion

Add tests for tracking data values

The action will vary based on the rendered button state. This will allow
us to separate these out for users that have opted into  tracking.

May as well put some checks on for the other required tracking values,
as things will break if they get accidentially removed.

Amend typos in test descriptions, switch to POST form