Clarify that you can only manage existing subscriptions here
We frequently get support tickets along the lines of "I've entered my
email address on this page a dozen times and got not emails!  Your
system is preventing me from signing up to notifications on XYZ!"

But the problem is that you can't sign up to new notifications from
this page.  The text on this page doesn't say that, and the text on
the next page says to contact support.  So, users who find this page
first get frustrated.

So, really emphasise that this is for *existing* subscriptions, not
new ones.