Add Feature Sign up with Accounts (#1123)
The provisional foundations for allowing sign up to a wider range of
notifications by content type, and enforcing the use of a GOV.UK Account
instead of creating a notifications specific account via email sign up.

The feature is currently gated behind a flag. We will need to add
finalised content before switching it on in production.

After this a follow up PR can remove code relating to the old sign up
routes.